Finish adding managers, owners person and owners company

+++ FinancialServices/Controllers/ManagerController.cs
@@ -11,9 +11,14 @@ namespace FinancialServices.Controllers
 
         private readonly IManagerService managerService;
 
-        public ManagerController(IManagerService _managerService)
+        private readonly IPersonService personService;
+
+        public ManagerController(
+            IManagerService _managerService,
+            IPersonService _personService)
         {
             managerService = _managerService;
+            personService = _personService;
 
         }
 
@@ -23,34 +28,46 @@ namespace FinancialServices.Controllers
         {
             var model = new AddManagerViewModel();
 
-
             return View(model);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Add(long idEik,AddManagerViewModel model)
+        public async Task<IActionResult> Add(long idEik, AddManagerViewModel model)
         {
             if (!ModelState.IsValid)
             {
                 return View(model);
             }
 
-            try
+
+            if (await personService.isPersonExist(model.IdEgn))
             {
-                await managerService.AddManagerAsync( idEik,model);
+                try
+                {
+                    await managerService.AddManagerAsync(idEik, model);
 
-                TempData[MessageConstants.SiccessMessage] = "Успешно добавихте мениджър";
+                    TempData[MessageConstants.SiccessMessage] = "Успешно добавихте мениджър";
 
-                return RedirectToAction("Details","Company", new { idEik = idEik });
+                    return RedirectToAction("Details", "Company", new { idEik = idEik });
+                }
+                catch (Exception)
+                {
+
+                    TempData[MessageConstants.WarningMessage] = "Вече сте добавичи Ф.Л. във фирмата";
+
+                    return View(model);
+
+                }
+               
+
             }
-            catch (Exception)
+            else
             {
-                ModelState.AddModelError("", "Something went wrong");
-
-                TempData[MessageConstants.WarningMessage] = "Липсва мениджър с това ЕГН във базата";
+                TempData[MessageConstants.WarningMessage] = "Няма мениджър с такова ЕГН";
 
                 return View(model);
             }
+
         }
 
 
@@ -58,7 +75,6 @@ namespace FinancialServices.Controllers
         public async Task<IActionResult> Delete(long idEik)
         {
             var model = new AddManagerViewModel();
-
 
             return View(model);
         }
@@ -71,24 +87,21 @@ namespace FinancialServices.Controllers
                 return View(model);
             }
 
-            try
+            if (await personService.isPersonExist(model.IdEgn) == true)
             {
                 await managerService.DeleteAsync(idEik, model.IdEgn);
 
                 TempData[MessageConstants.SiccessMessage] = "Успешно изтрихте мениджър";
 
-
-
                 return RedirectToAction("Details", "Company", new { idEik = idEik });
             }
-            catch (Exception)
+            else
             {
-                ModelState.AddModelError("", "Something went wrong");
-
-                TempData[MessageConstants.WarningMessage] = "Липсва мениджър с това ЕГН във базата";
+                TempData[MessageConstants.WarningMessage] = "Няма човек стакова егн";
 
                 return View(model);
             }
+
         }
 
     }
